requirements.txt
# requirements.txt

# Core Scientific Libraries
numpy==1.23.5          # Fundamental package for numerical computation
scipy==1.10.1          # Library for scientific and technical computing (e.g., signal processing)
matplotlib==3.6.0      # Plotting and visualization
pandas==1.5.3          # Data manipulation and analysis

# Libraries for Electrophysiological Analysis
elephant==0.11.0       # Electrophysiology data analysis toolkit (spike train analysis)
neo==0.10.2            # Handling electrophysiological data (Neo object support)
pyabf==2.3.5           # For handling ABF (Axon Binary Format) files
spikeinterface==0.98.2 # Unified framework for spike sorting
quantities==0.14.1     # Physical quantities with units

# Libraries for Time-Frequency Analysis
pywavelets==1.4.1      # Wavelet transforms for time-frequency analysis

# Libraries for Network and Connectivity Analysis
networkx==3.1          # Network analysis and graph theory

# Statistical Modeling and Curve Fitting
statsmodels==0.14.0    # Statistical models for dose-response analysis and more

# Machine Learning and Clustering
scikit-learn==1.3.0    # Machine learning library
hdbscan==0.8.33        # Hierarchical density-based clustering
umap-learn==0.5.4      # Uniform Manifold Approximation and Projection

# Visualization Libraries
plotly==5.17.0         # Interactive plotting library
datashader==0.15.2     # Large dataset visualization

# Utilities
pyyaml==6.0.1          # YAML file parsing
joblib==1.3.2          # Parallel processing utilities